+++ src/main/java/com/project/easywork/client/domain/dto/stack_measurement/StackMeasurementUpdateD.java
@@ -5,10 +5,12 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 public class StackMeasurementUpdateD {
   private Cycle cycle;
-  private Double allowance;
+  private BigDecimal allowance;
 }
